+++ Source/DKMethod.m
@@ -398,7 +398,7 @@
    */
 
   // Initial type string containing self and _cmd.
-  NSMutableString *typeString = [[NSMutableString alloc] initWithFormat: @"@0:%lu", (unsigned long)sizeof(id)];
+  NSMutableString *typeString = [[NSMutableString alloc] initWithFormat: @"@0:%"PRIuPTR"", sizeof(id)];
   NSUInteger offset = sizeof(id) + sizeof(SEL);
   NSString *returnValue = nil;
   NSEnumerator *en = [inArgs objectEnumerator];
@@ -416,7 +416,7 @@
       typeChar = [arg unboxedObjCTypeChar];
     }
 
-    [typeString appendFormat: @"%s%lu", typeChar, (unsigned long)offset];
+    [typeString appendFormat: @"%s%"PRIuPTR"", typeChar, offset];
 
     if (doBox)
     {
@@ -428,7 +428,7 @@
     }
   }
 
-  returnValue = [NSString stringWithFormat: @"%s%lu%@", [self returnTypeBoxed: doBox],
+  returnValue = [NSString stringWithFormat: @"%s%"PRIuPTR"%@", [self returnTypeBoxed: doBox],
     (unsigned long)offset,
     typeString];
   [typeString release];
